refactor(file): Move FileAction to its own file.
